requirements.txt
streamlit>=1.29.0
langchain>=0.1.0
langchain-anthropic>=0.1.1
networkx>=3.2.1
pyvis>=0.3.2
python-dotenv>=1.0.0
beautifulsoup4>=4.12.2
requests>=2.31.0
matplotlib>=3.8.2
lxml>=4.9.0
neo4j>=5.14.0